+++ src/agents/policy-manager.ts
@@ -7,35 +7,38 @@ import {
 } from '../../types/ethers-contracts';
 import { TransactingPower } from '../crypto/powers';
 import { ChecksumAddress } from '../types';
+import { toHexString } from '../utils';
 
 import { CONTRACTS, DEFAULT_WAIT_N_CONFIRMATIONS } from './constants';
 
 export class PolicyManagerAgent {
   public static async createPolicy(
-    policyId: Uint8Array,
     transactingPower: TransactingPower,
+    policyId: Uint8Array,
     valueInWei: number,
     expirationTimestamp: number,
     nodeAddresses: Array<ChecksumAddress>,
-    ownerAddress?: ChecksumAddress
+    ownerAddress: ChecksumAddress
   ): Promise<ContractTransaction> {
-    const PolicyManager = await this.connect(transactingPower.signer);
-
+    const PolicyManager = await this.connect(
+      transactingPower.provider,
+      transactingPower.signer
+    );
     // TODO: Call fails due to "UNPREDICTABLE_GAS_LIMIT" error, hard-coding `gasLimit` for now
     // const estimatedGas = await PolicyManager.estimateGas.createPolicy(
     //   policyId,
-    //   ownerAddress ?? transactingPower.account,
+    //   ownerAddress,
     //   expirationTimestamp,
     //   nodeAddresses
     // );
     const overrides = {
       // gasLimit: estimatedGas.toNumber(),
-      gasLimit: 200_000,
-      value: valueInWei,
+      gasLimit: 350_000,
+      value: BigInt(valueInWei),
     };
     const tx = await PolicyManager.createPolicy(
       hexlify(policyId),
-      ownerAddress ?? (await transactingPower.signer.getAddress()),
+      ownerAddress,
       expirationTimestamp,
       nodeAddresses,
       overrides
@@ -44,11 +47,41 @@ export class PolicyManagerAgent {
     return tx;
   }
 
+  public static async policyDisabled(
+    provider: ethers.providers.Provider,
+    policyId: Uint8Array
+  ): Promise<boolean> {
+    const PolicyManager = await this.connect(provider);
+    const policy = await PolicyManager.policies(policyId);
+    if (!policy) {
+      throw Error(`Policy with id ${toHexString(policyId)} does not exist.`);
+    }
+    return policy[0];
+  }
+
+  public static async revokePolicy(
+    transactingPower: TransactingPower,
+    policyId: Uint8Array
+  ): Promise<ContractTransaction> {
+    const PolicyManager = await this.connect(
+      transactingPower.provider,
+      transactingPower.signer
+    );
+    const estimatedGas = await PolicyManager.estimateGas.revokePolicy(policyId);
+    const overrides = {
+      gasLimit: estimatedGas.toNumber(),
+    };
+    const tx = await PolicyManager.revokePolicy(hexlify(policyId), overrides);
+    await tx.wait(DEFAULT_WAIT_N_CONFIRMATIONS);
+    return tx;
+  }
+
   private static async connect(
-    signer: ethers.providers.JsonRpcSigner
+    provider: ethers.providers.Provider,
+    signer?: ethers.providers.JsonRpcSigner
   ): Promise<PolicyManager> {
-    const network = await signer.provider.getNetwork();
+    const network = await provider.getNetwork();
     const contractAddress = CONTRACTS[network.name].POLICYMANAGER;
-    return PolicyManager__factory.connect(contractAddress, signer);
+    return PolicyManager__factory.connect(contractAddress, signer ?? provider);
   }
 }